Move fragment source files to another package.
